added security to the views

+++ products/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.views import generic, View
 from django.db.models.functions import Lower
 
@@ -90,6 +91,7 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
+@login_required
 def toggle_favorite(request, product_id):
     product = get_object_or_404(Product, pk=product_id)
 
@@ -102,6 +104,7 @@ def toggle_favorite(request, product_id):
 
     return redirect(request.META['HTTP_REFERER'])
 
+@login_required
 def favorite_products(request):
     """" 
     retrieves a list of favorite products 
@@ -122,8 +125,15 @@ def favorite_products(request):
     
     return render(request, 'products/favorite_products.html', context)
 
+@login_required
 def add_product(request):
     """ Add products to the shop """
+
+    """ Validate Super User """
+    if not request.user.is_superuser:
+        messages.error(request, 'Apologies, this action is limited to store owners only.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -142,8 +152,15 @@ def add_product(request):
 
     return render(request, template, context)
 
+@login_required
 def edit_product(request, product_id):
     """ Update the shop products """
+    
+    """ Validate Super User """
+    if not request.user.is_superuser:
+        messages.error(request, 'Apologies, this action is limited to store owners only.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -165,8 +182,15 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+
+    """ Validate Super User """
+    if not request.user.is_superuser:
+        messages.error(request, 'Apologies, this action is limited to store owners only.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
